Switch Music Playback Backend to Kiri
Previously, LRCGET utilized the howler.js library for audio playback,
which in turn relied on the platform's webview engine. This approach led
to inconsistent audio support across different platforms (Windows,
Linux, macOS), along with several platform-specific bugs that required
complex (nasty) workarounds, not to mention the insecure file server hack!

In this commit, I have completely removed the dependency on howler.js
and transitioned to using Kiri, a frontend for the Symphonia library.
Both Kiri and Symphonia are native Rust libraries, which are expected to
provide consistent audio support across platforms for LRCGET. A big
thank you to all the developers involved in these projects!