Fix issue with saving unsynced lyrics as synced
This commit resolves the problem of unsynced lyrics incorrectly being
displayed as synced lyrics. The implemented solution involves detecting
the presence of at least one pair of brackets `[]` at the start of any
line. If such a pair exists, the lyrics are classified as synced and
saved as an .lrc file. Otherwise, the lyrics are saved as a .txt file
and the track are displayed as having plain lyrics in track list.