Seek: round progressPercent to thousandths
The `progress` currently updates every `40ms` which re-computes
`progressPercent`. Vue needs to do a layout each time in order to update
the `ViewSliderDot` inside `Seek`'s `VueSlider`.

To reduce the CPU usage of these updates and make the interface more
responsive, round `progressPercent` to the nearest `0.001`. This will
limit layout updates to about `1000` over the duration of a track. Since
this is approximately the same as the default window width, there isn't
much impact to the smoothness of the seek slider while playing a track.
If the user resizes the window to be wider, the slider will scroll slightly
less smoothly than before. However, this does not reduce the precision if
the user seeks manually.