TrackItem: hide Equalizer when editing lyrics
The style animations used by the equalizer are surprisingly expensive.
It's a nice visual feedback to show which track is playing, but it
impacts the user experience when editing lyrics. Since the `EditLyrics`
modal covers most of the track list anyways, hide the `Equalizer` when
the modal is open to avoid CPU impact from the "Styles Invalidated".