linux: add desktop file
This also installs a desktop file in the DEB and RPM packages.

It has hardcoded values on purpose, to make it easier to use it with
packaging formats not supported by Tauri (such as Flatpak).

This desktop file is not installed in the AppImage package, because
Tauri's AppImageConfig doesn't support the "desktopTemplate" option,
like DEB and RPM packages do.

Docs: https://specifications.freedesktop.org/desktop-entry-spec/latest/recognized-keys.html